[refactor] use bump instead of conversion as we write a conversion to migrate from spdx2 to spdx3
