+++ tests/spdx3/bump/test_snippet_bump.py
@@ -8,3 +8,18 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
+from unittest import mock
+
+from spdx3.bump_from_spdx2.snippet import bump_snippet
+from tests.fixtures import snippet_fixture
+from spdx.model.snippet import Snippet as Spdx2_Snippet
+from spdx3.model.software.snippet import Snippet
+
+
+@mock.patch("spdx3.model.creation_information.CreationInformation", autospec=True)
+def test_bump_snippet(creation_information):
+    spdx2_snippet: Spdx2_Snippet = snippet_fixture()
+
+    snippet: Snippet = bump_snippet(spdx2_snippet, creation_information=creation_information)
+
+    assert snippet.spdx_id == "SPDXRef-Snippet"
